backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai>=0.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
protobuf>=4.25.0,<5.0.0
opentelemetry-proto<1.34.0
psycopg2-binary>=2.9.0
python-json-logger==2.0.7
